+++ src/services/DemoRunner.ts
@@ -38,6 +38,8 @@ import {
   getUserInput,
   clearVariablesState,
   setContext,
+  writeFile,
+  applyPatch,
 } from "../utils";
 import { ActionTreeItem } from "../providers/ActionTreeviewProvider";
 import { DecoratorService } from "./DecoratorService";
@@ -646,7 +648,12 @@ export class DemoRunner {
       }
 
       if (step.action === Action.Create) {
-        await workspace.fs.writeFile(fileUri, new Uint8Array(Buffer.from(content)));
+        await writeFile(fileUri, content);
+        continue;
+      }
+
+      if (step.action === Action.ApplyPatch) {
+        await applyPatch(fileUri, content, step.patch);
         continue;
       }