+++ src/utils/writeFile.ts
@@ -1,0 +1,5 @@
+import { Uri, workspace } from "vscode";
+
+export const writeFile = async (filePath: Uri, text: string) => {
+  await workspace.fs.writeFile(filePath, new TextEncoder().encode(text));
+};